+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Armor rebalance now supports armors from Immersive Armors.
+- Fixed missing attribute text and odd formatting on the tooltips of armor items enchanted with protection enchantments when "Enchantment-based toughness" is on.